Trait gear_lib::fungible_token::ft_core::FTCore

pub trait FTCore: FTStateKeeper {
    // Provided methods
    fn mint(&mut self, to: &ActorId, amount: u128) { ... }
    fn burn(&mut self, amount: u128) { ... }
    fn transfer(&mut self, from: &ActorId, to: &ActorId, amount: u128) { ... }
    fn approve(&mut self, to: &ActorId, amount: u128) { ... }
    fn can_transfer(&mut self, from: &ActorId, amount: u128) -> bool { ... }
}

Provided Methods

fn mint(&mut self, to: &ActorId, amount: u128)

Mints amount of token

Requirements: None Arguments:

amount: The amount of token to be minted (actually have no limit)

fn burn(&mut self, amount: u128)

Burns amount of token

Requirements:

msg::source() MUST have enough tokens on his balance Arguments: amount: The amount of token to be burnt

fn transfer(&mut self, from: &ActorId, to: &ActorId, amount: u128)

Transfer amount of token

Requirements:

Only the token owner or approved account can call that action

from MUST have enough tokens

from and to MUST be non-zero addresses

Arguments:

from: An account from which token will be transerred

to: An account to which token will be transferred

amount: The amount of token of be transferred

fn approve(&mut self, to: &ActorId, amount: u128)

Gives a right to another account to manage the amount of token

Requirements:

Only the token owner can call that action

to MUST be a non-zero account

Arguments:

to: An account that will be approved to manage the indicated amount of token

amount: The amount of tokens to be approved

fn can_transfer(&mut self, from: &ActorId, amount: u128) -> bool

Checks whether it is possible to perform a transfer