Trait gear_lib::multitoken::mtk_core::MTKCore

pub trait MTKCore: StateKeeper + MTKTokenState {
    // Provided methods
    fn assert_can_burn(&mut self, owner: &ActorId, id: &TokenId, amount: u128) { ... }
    fn assert_can_transfer(&self, from: &ActorId, id: &u128, amount: u128) { ... }
    fn assert_approved(&self, owner: &ActorId, operator: &ActorId) { ... }
    fn mint_impl(
        &mut self,
        account: &ActorId,
        id: &TokenId,
        amount: u128,
        meta: Option<TokenMetadata>
    ) { ... }
    fn mint(
        &mut self,
        account: &ActorId,
        ids: Vec<TokenId>,
        amounts: Vec<u128>,
        meta: Vec<Option<TokenMetadata>>
    ) { ... }
    fn burn_impl(&mut self, id: &TokenId, amount: u128) { ... }
    fn burn(&mut self, ids: Vec<TokenId>, amounts: Vec<u128>) { ... }
    fn transfer_from_impl(
        &mut self,
        from: &ActorId,
        to: &ActorId,
        id: &TokenId,
        amount: u128
    ) { ... }
    fn transfer_from(
        &mut self,
        from: &ActorId,
        to: &ActorId,
        ids: Vec<TokenId>,
        amounts: Vec<u128>
    ) { ... }
    fn approve(&mut self, to: &ActorId) { ... }
    fn revoke_approval(&mut self, to: &ActorId) { ... }
    fn balance_of(&self, accounts: Vec<ActorId>, ids: Vec<TokenId>) { ... }
}

Provided Methods

fn assert_can_burn(&mut self, owner: &ActorId, id: &TokenId, amount: u128)

fn assert_can_transfer(&self, from: &ActorId, id: &u128, amount: u128)

fn assert_approved(&self, owner: &ActorId, operator: &ActorId)

fn mint_impl( &mut self, account: &ActorId, id: &TokenId, amount: u128, meta: Option<TokenMetadata> )

fn mint( &mut self, account: &ActorId, ids: Vec<TokenId>, amounts: Vec<u128>, meta: Vec<Option<TokenMetadata>> )

Mints multiple new tokens (in case all input length is 1 - simple mint) Requirements:

ids element must a unique value

account must be a non-zero account Arguments:

account: An account to which minted token will be assigned

ids: The vector of IDs of minted tokens

amounts: The vector of amounts of tokens to mint (1 in case of an NFT)

meta: The vector of optional additional metadata for NFTs

fn burn_impl(&mut self, id: &TokenId, amount: u128)

fn burn(&mut self, ids: Vec<TokenId>, amounts: Vec<u128>)

Burns multiple tokens (in case all input length is 1 - simple burn) Requirements:

Only token owner can perform this action

ids element must be the ID of the existing token

amounts element must not exceed user’s token balance Arguments:

ids: The vector of ids of the token to be burnt

amounts: The vector of amounts of token to be burnt

fn transfer_from_impl( &mut self, from: &ActorId, to: &ActorId, id: &TokenId, amount: u128 )

fn transfer_from( &mut self, from: &ActorId, to: &ActorId, ids: Vec<TokenId>, amounts: Vec<u128> )

Transfers multiple tokens to a new user (in case all input length is 1 - simple transfer) Requirements:

Only the token owner or approved account can call that action

to must be a non-zero account

ids element must be the ID of the existing token

amounts element must not exceed from’s balance Arguments:

from: An account from which token will be transferred

to: An account to which token will be transferred

ids: The vector of IDs of transferred token

amounts: The vector of amounts of transferred token

fn approve(&mut self, to: &ActorId)

Gives a right to another account to manage its tokens Requirements:

Only the token owner can call that action

to must be a non-zero account Arguments:

to: An account that will be approved to manage the tokens

fn revoke_approval(&mut self, to: &ActorId)

Removed a right to another account to manage its tokens Requirements:

Only the token owner can call that action

to must be a non-zero account Arguments:

to: An account that won’t be able to manage the tokens

fn balance_of(&self, accounts: Vec<ActorId>, ids: Vec<TokenId>)

Returns the amount of multiple specific tokens multiple users have (in case all input length is 1 - simple balance_of) Arguments:

accounts: The vectors of IDs of the actor

id: The vector of token IDs which balance will be returned