Trait gear_lib::non_fungible_token::nft_core::NFTCore

pub trait NFTCore: NFTStateKeeper {

    fn mint(
        &mut self,
        to: &ActorId,
        token_id: TokenId,
        token_metadata: Option<TokenMetadata>
    ) -> NFTTransfer { ... }
    fn burn(&mut self, token_id: TokenId) -> NFTTransfer { ... }
    fn transfer(&mut self, to: &ActorId, token_id: TokenId) -> NFTTransfer { ... }
    fn transfer_payout(
        &mut self,
        to: &ActorId,
        token_id: TokenId,
        amount: u128
    ) -> NFTTransferPayout { ... }
    fn internal_transfer(&mut self, to: &ActorId, token_id: TokenId) -> ActorId { ... }
    fn approve(&mut self, to: &ActorId, token_id: TokenId) -> NFTApproval { ... }
    fn revoke_approval(
        &mut self,
        approved_account: &ActorId,
        token_id: TokenId
    ) -> NFTApproval { ... }
    fn owner_of(&self, token_id: TokenId) -> ActorId { ... }
    fn is_approved_to(&self, to: &ActorId, token_id: TokenId) -> bool { ... }
    fn delegated_approve(
        &mut self,
        message: DelegatedApproveMessage,
        signed_approve: [u8; 64]
    ) -> NFTApproval { ... }
    fn nft_payout(&self, owner: &ActorId, amount: u128) -> Payout { ... }
    fn assert_token_exists(&self, token_id: TokenId) { ... }
    fn assert_zero_address(&self, account: &ActorId) { ... }
    fn assert_can_transfer(&self, token_id: TokenId, owner: &ActorId) { ... }
    fn assert_owner(&self, owner: &ActorId) { ... }

}

Provided Methods

fn mint(
&mut self,
to: &ActorId,
token_id: TokenId,
token_metadata: Option<TokenMetadata>
) -> NFTTransfer

Mints a new token

Requirements:

token_id must be unique

to must be a non-zero account

Arguments:

to: An account to which minted NFT will be assigned

token_id: the ID of minted NFT

token_metadata: optional additional metadata about NFT

fn burn(&mut self, token_id: TokenId) -> NFTTransfer

Burns a token

Requirements:

Only NFT owner can call that action

token_id must be the ID of the existing NFT

Arguments:

token_id: the ID of NFT that will be burnt

fn transfer(&mut self, to: &ActorId, token_id: TokenId) -> NFTTransfer

Transfers a token to the new owner

Requirements:

Only the token owner or approved account can call that action

to must be a non-zero account

token_id must be the ID of the existing NFT

Arguments:

to: An account to which NFT will be transferred

token_id: the ID of transferred NFT

fn transfer_payout(
&mut self,
to: &ActorId,
token_id: TokenId,
amount: u128
) -> NFTTransferPayout

Transfers a token to the new owner

Requirements:

Only the token owner or approved account can call that action

to must be a non-zero account

token_id must be the ID of the existing NFT

Arguments:

to: An account to which NFT will be transferred

token_id: the ID of transferred NFT

fn internal_transfer(&mut self, to: &ActorId, token_id: TokenId) -> ActorId

fn approve(&mut self, to: &ActorId, token_id: TokenId) -> NFTApproval

Gives a right to another account to manage the token with indicated ID

Requirements:

Only the token owner can call that action

to must be a non-zero account

token_id must be the ID of the existing NFT

Arguments:

to: An account that will be approved to manage the indicated NFT

token_id: the ID of the NFT

fn revoke_approval(
&mut self,
approved_account: &ActorId,
token_id: TokenId
) -> NFTApproval

fn owner_of(&self, token_id: TokenId) -> ActorId

fn is_approved_to(&self, to: &ActorId, token_id: TokenId) -> bool

fn delegated_approve(
&mut self,
message: DelegatedApproveMessage,
signed_approve: [u8; 64]
) -> NFTApproval

fn nft_payout(&self, owner: &ActorId, amount: u128) -> Payout

Returns a Payout struct for a given token If NFT contract has no royalties it just returns BtreeMap {“owner”: “amount”}

fn assert_token_exists(&self, token_id: TokenId)

Checks that NFT with indicated ID already exists

fn assert_zero_address(&self, account: &ActorId)

Checks account for a zero address

fn assert_can_transfer(&self, token_id: TokenId, owner: &ActorId)

Checks that msg::source() is allowed to manage the token with indicated token_id

fn assert_owner(&self, owner: &ActorId)

Checks that msg::source() is the owner of the token with indicated token_id